print numbers

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -7,7 +7,7 @@
 void print_number(int n)
 {
 	int m; /* power of 10 */
-	int c; /* boolean check */
+	float c; /* boolean check */
 	int num;
 
 	num = n;
